add : sql

+++ la-mia-pizzeria/src/main/java/com/esercizio/pizzeria/repository/PizzaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.esercizio.pizzeria.Model.Pizza;
 
-public interface PizzaRepository extends JpaRepository<Pizza, Long> {
+public interface PizzaRepository extends JpaRepository<Pizza, Integer> {
 
 }
